requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
flask-ngrok==0.0.25
matplotlib==3.2.2
numpy==1.18.5
Pillow==7.1.2
python-dotenv==0.14.0
tensorflow==2.2.0
tensorflow-hub==0.7.0
Werkzeug==1.0.1